Extract SimpleMiniPlayerView and make MiniPlayerViewController generic class
